+++ CHANGELOG.md
@@ -24,6 +24,9 @@
 
 * Fix a number of `@extend` bugs:
 
+  * `selector-extend()` and `selector-replace()` now allow compound selector
+    extendees.
+
   * Remove the universal selector `*` when unifying with other selectors.
 
   * Properly unify the result of multiple simple selectors in the same compound
@@ -37,6 +40,8 @@
     [second law of `@extend`][laws].
 
   * Allow extensions that match selectors but fail to unify.
+
+  * Partially-extended selectors are no longer used as parent selectors.
 
   * Fix an edge case where both the extender and the extended selector
     have invalid combinator sequences.